add : commentForm에 프로필사진 추가

+++ src/main/java/com/botox/service/RoomService.java
@@ -124,11 +124,13 @@ public class RoomService {
     @Transactional
     public void checkAndDeleteEmptyRooms() {
         // UserCount가 0인 방을 찾는 쿼리를 실행합니다.
-        List<Room> emptyRooms = roomRepository.findRoomsByUserCount(0);
+        List<Room> emptyRooms = roomRepository.findByUserCount(0);
 
         // 찾은 빈 방들을 삭제합니다.
         for (Room room : emptyRooms) {
-            roomRepository.delete(room);
+            if(room.getRoomMaster()== null) {
+                roomRepository.delete(room);
+            }
         }
     }
     // 방 나가기 기능
